Add back conda environment file for docs only
This is not ideal, as this adds redundancy between the `requirements-dev.txt` file and the `docs/env.yml` file, but the build environment for ReadTheDocs is limited in its capability.  We cannot create the necessary MPI environment without conda, and we cannot mix conda and regular virtualenvs.  So, we need to create a full conda environment file that duplicates some of the dependencies in the `requirements-dev.txt` file.